requirements.txt
anthropic>=0.39.0
openai>=1.50.0
pydantic>=2.0.0
pyyaml>=6.0
yfinance>=0.2.30
tabulate>=0.9.0
